add curl and teletone to mod_spidermonkey.  fix /tmp/ dir for windows, added global temp dir var.  Multiple type fixes.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@713 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/include/switch_ivr.h
@@ -140,7 +140,7 @@ SWITCH_DECLARE(switch_status) switch_ivr_speak_text(switch_core_session *session
 													char *tts_name,
 													char *voice_name,
 													char *timer_name,
-													unsigned int rate,
+													size_t rate,
 													switch_dtmf_callback_function dtmf_callback,
 													char *text,
 													void *buf,
